Move check logic from utils.c to checks.c
There where "utils" in utils.c to handle checks, which was just only
check logic. Those should be part of the checks module, so we don't
have a dependency from utils.c to checks.c/h
