+++ naemon/checks.h
@@ -33,6 +33,13 @@ void checks_init(void); /* Init check execution, schedule events */
 int parse_check_output(char *, char **, char **, char **, int, int);
 struct check_output *parse_output(const char *, struct check_output *);
 
+int process_check_result_queue(char *);
+int process_check_result_file(char *);
+int process_check_result(check_result *);
+int delete_check_result_file(char *);
+int init_check_result(check_result *);
+int free_check_result(check_result *);                  	/* frees memory associated with a host/service check result */
+
 NAGIOS_END_DECL
 
 #endif
